Move ConfirmModal and Summary components to TemporaryPage
