+++ src/App.scss
@@ -15,13 +15,3 @@ html {
 a {
   text-decoration: none;
 }
-
-.main {
-  min-height: 100vh;
-
-  &__container {
-    display: flex;
-    gap: 2rem;
-    margin: 2rem 0 0 2rem;
-  }
-}
