+++ src/pages/TemporaryPage/TemporaryPage.jsx
@@ -1,0 +1,32 @@
+import { useState } from "react";
+import "./TemporaryPage.scss";
+import Temp from "../../components/ConfirmModal/Temp";
+import Summary from "../../components/Summary/Summary";
+
+function TemporaryPage() {
+  const [contribution, setContribution] = useState(1200.0);
+
+  const handleContributionChange = (increase) => {
+    setContribution(contribution + increase);
+  };
+
+  return (
+    <main className="main">
+      <div className="main__container">
+        <Summary
+          contribution={1000}
+          theRest={20010.33}
+          initiative="Better World Initiative"
+        />
+        <Summary
+          contribution={contribution}
+          theRest={10093.59}
+          initiative="Some Other Initiative"
+        />
+      </div>
+      <Temp donate={handleContributionChange} />
+    </main>
+  );
+}
+
+export default TemporaryPage;
